refactor: update autocompleteUnstyled folder to useAutocomplete.
