+++ docs/data/material/components/autocomplete/CustomizedHook.tsx
@@ -1,8 +1,5 @@
 import * as React from 'react';
-import {
-  useAutocomplete,
-  AutocompleteGetTagProps,
-} from '@mui/base/AutocompleteUnstyled';
+import { useAutocomplete, AutocompleteGetTagProps } from '@mui/base/useAutocomplete';
 import CheckIcon from '@mui/icons-material/Check';
 import CloseIcon from '@mui/icons-material/Close';
 import { styled } from '@mui/material/styles';
