+++ packages/mui-base/src/useAutocomplete/useAutocomplete.spec.ts
@@ -1,4 +1,4 @@
-import { useAutocomplete, FilterOptionsState } from '@mui/base/AutocompleteUnstyled';
+import { useAutocomplete, FilterOptionsState } from '@mui/base/useAutocomplete';
 import { expectType } from '@mui/types';
 
 interface Person {
