+++ packages/mui-base/src/useAutocomplete/useAutocomplete.test.js
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { expect } from 'chai';
 import { createRenderer, screen, ErrorBoundary, act, fireEvent } from 'test/utils';
-import { useAutocomplete, createFilterOptions } from '@mui/base/AutocompleteUnstyled';
+import { useAutocomplete, createFilterOptions } from '@mui/base/useAutocomplete';
 
 describe('useAutocomplete', () => {
   const { render } = createRenderer();
