+++ packages/mui-material/src/useAutocomplete/useAutocomplete.js
@@ -1,2 +1,2 @@
-export { useAutocomplete as default } from '@mui/base/AutocompleteUnstyled';
-export * from '@mui/base/AutocompleteUnstyled';
+export { useAutocomplete as default } from '@mui/base/useAutocomplete';
+export * from '@mui/base/useAutocomplete';
